Cleanup acceptance tests
Also fixed the sorting issue with Firefox and problem introduced in #396 casuing travis acceptance tests to always fail
