+++ acj/static/test/features/edit_assignment.feature
@@ -2,19 +2,19 @@ Feature: Edit Assignment
   As user, I want to edit an assignment
 
   Scenario: Loading edit assignment page as admin
-    Given I'm a System Administrator with assignments
+    Given I'm a System Administrator
     And I'm on 'assignment' page for assignment with id '1abcABC123-abcABC123_Z' and course id '1abcABC123-abcABC123_Z'
     When I select 'Edit Assignment' button
     Then I should be on the 'edit assignment' page
 
   Scenario: Loading edit assignment page as instructor
-    Given I'm an Instructor with assignments
+    Given I'm an Instructor
     And I'm on 'assignment' page for assignment with id '1abcABC123-abcABC123_Z' and course id '1abcABC123-abcABC123_Z'
     When I select 'Edit Assignment' button
     Then I should be on the 'edit assignment' page
 
   Scenario: Editing an assignment as instructor
-    Given I'm an Instructor with assignments
+    Given I'm an Instructor
     And I'm on 'edit assignment' page for assignment with id '3abcABC123-abcABC123_Z' and course id '1abcABC123-abcABC123_Z'
     When I fill form item 'assignment.name' in with 'New Name'
     And I fill in the assignment description with 'This is the new description'
@@ -25,7 +25,7 @@ Feature: Edit Assignment
     And I should see the assignment with the new name and description
 
   Scenario: Editing a assignment's criterion as instructor before comparisons
-    Given I'm an Instructor with assignments
+    Given I'm an Instructor
     And I'm on 'edit assignment' page for assignment with id '3abcABC123-abcABC123_Z' and course id '1abcABC123-abcABC123_Z'
     When I edit the second criterion
     And I fill form item 'criterion.name' in with 'Choose the best one'
@@ -40,13 +40,13 @@ Feature: Edit Assignment
     Then I should be on the 'course' page
 
   Scenario: Editing assignment's criterion as instructor after the criterion has been compared
-    Given I'm an Instructor with assignments
+    Given I'm an Instructor
     And I'm on 'edit assignment' page for assignment with id '1abcABC123-abcABC123_Z' and course id '1abcABC123-abcABC123_Z'
     When I edit the second criterion
     Then I should see a warning message in the edit criterion modal
 
   Scenario: Cannot add or remove an assignment's criterion as instructor after its been compared
-    Given I'm an Instructor with assignments
+    Given I'm an Instructor
     And I'm on 'edit assignment' page for assignment with id '1abcABC123-abcABC123_Z' and course id '1abcABC123-abcABC123_Z'
     Then I should not be able to add criteria
     And I should not be able to remove criteria
