+++ acj/static/test/features/edit_course.feature
@@ -2,19 +2,19 @@ Feature: Edit Course
   As user, I want to edit a course
 
   Scenario: Loading edit course page as admin
-    Given I'm a System Administrator with courses
+    Given I'm a System Administrator
     And I'm on 'course' page for course with id '1abcABC123-abcABC123_Z'
     When I select 'Edit Course' button
     Then I should be on the 'edit course' page
 
   Scenario: Loading edit course page as instructor
-    Given I'm an Instructor with courses
+    Given I'm an Instructor
     And I'm on 'course' page for course with id '1abcABC123-abcABC123_Z'
     When I select 'Edit Course' button
     Then I should be on the 'edit course' page
 
   Scenario: Editing a course as instructor
-    Given I'm an Instructor with courses
+    Given I'm an Instructor
     And I'm on 'edit course' page for course with id '1abcABC123-abcABC123_Z'
     When I fill form item 'course.name' in with 'New Name'
     And I fill form item 'course.year' in with '2020'
