+++ acj/static/test/features/edit_course_user.feature
@@ -2,13 +2,13 @@ Feature: Edit Course Users
   As user, I want to manage the users of a course
 
   Scenario: Loading manage course users page as admin
-    Given I'm a System Administrator with courses
+    Given I'm a System Administrator
     And I'm on 'course' page for course with id '1abcABC123-abcABC123_Z'
     When I select 'Manage Users' button
     Then I should be on the 'manage users' page
 
   Scenario: Loading manage course users page as instructor
-    Given I'm an Instructor with students
+    Given I'm an Instructor
     And I'm on 'course' page for course with id '1abcABC123-abcABC123_Z'
     When I select 'Manage Users' button
     Then I should be on the 'manage users' page
@@ -18,8 +18,8 @@ Feature: Edit Course Users
       | First Instructor      |
       | First Student         |
 
-  Scenario: Sorting course user as instructor
-    Given I'm an Instructor with students
+  Scenario: Sorting course users as instructor
+    Given I'm an Instructor
     And I'm on 'edit course user' page for course with id '1abcABC123-abcABC123_Z'
     When I sort by displayname in decending order
     Then I should see course users with displaynames:
@@ -28,7 +28,7 @@ Feature: Edit Course Users
       | First Instructor       |
 
   Scenario: Adding user to course as instructor
-    Given I'm an Instructor with students
+    Given I'm an Instructor
     And I'm on 'edit course user' page for course with id '1abcABC123-abcABC123_Z'
     When I fill form item 'user' in with 'Second'
     And I select the first user search result
@@ -41,8 +41,8 @@ Feature: Edit Course Users
       | First Student          |
       | Second Student         |
 
-  Scenario: Removing user to course as instructor
-    Given I'm an Instructor with students
+  Scenario: Removing user from course as instructor
+    Given I'm an Instructor
     And I'm on 'edit course user' page for course with id '1abcABC123-abcABC123_Z'
     When I drop the second user from the course
     Then I should see '1' users listed for the course
@@ -51,13 +51,13 @@ Feature: Edit Course Users
       | First Instructor       |
 
   Scenario: Changing user's group in course as instructor
-    Given I'm an Instructor with students
+    Given I'm an Instructor
     And I'm on 'edit course user' page for course with id '1abcABC123-abcABC123_Z'
     When I set the second user's group to 'Second Group'
     Then I should see a success message
 
   Scenario: Removing user from group in course as instructor
-    Given I'm an Instructor with students
+    Given I'm an Instructor
     And I'm on 'edit course user' page for course with id '1abcABC123-abcABC123_Z'
     When I set the second user's group to '- None -'
     Then I should see a success message
