+++ acj/static/test/features/step_definitions/create_assignment.js
@@ -9,6 +9,7 @@ var expect = chai.expect;
 var createAssignmentStepDefinitionsWrapper = function () {
     this.Given("I fill in the assignment description with '$text'", function(text) {
         //load the ckeditor iframe
+        browser.sleep(1000);
         browser.wait(browser.isElementPresent(element(by.css("#cke_assignmentDescription iframe"))), 1000);
         browser.driver.switchTo().frame(element(by.css("#cke_assignmentDescription iframe")).getWebElement());
         // clear the content
