+++ acj/static/test/features/step_definitions/create_course.js
@@ -9,6 +9,7 @@ var expect = chai.expect;
 var createCourseStepDefinitionsWrapper = function () {
     this.Given("I fill in the course description with '$text'", function(text) {
         //load the ckeditor iframe
+        browser.sleep(1000);
         browser.wait(browser.isElementPresent(element(by.css("#cke_courseDescription iframe"))), 1000);
         browser.driver.switchTo().frame(element(by.css("#cke_courseDescription iframe")).getWebElement());
         // clear the content
