--- acj/static/test/features/step_definitions/setup_instructor.js
@@ -1,47 +1,0 @@
-// Use the external Chai As Promised to deal with resolving promises in
-// expectations
-var chai = require('chai');
-var chaiAsPromised = require('chai-as-promised');
-chai.use(chaiAsPromised);
-
-var expect = chai.expect;
-
-var PageFactory  = require('../../factories/page_factory.js');
-var backEndMocks = require('../../factories/http_backend_mocks.js');
-
-var setupInstructorStepDefinitionsWrapper = function () {
-    var pageFactory = new PageFactory();
-    var loginDialog = pageFactory.createPage('login');
-
-    this.Given("I'm an Instructor", {timeout: 20 * 1000}, function () {
-        var fixtureName = 'instructor/default_fixture';
-        backEndMocks.setStorageFixture(browser, fixtureName);
-        return loginDialog.login(backEndMocks.getLoginDetails(fixtureName));
-    });
-
-    this.Given("I'm an Instructor with students", {timeout: 20 * 1000}, function () {
-        var fixtureName = 'instructor/has_students_fixture';
-        backEndMocks.setStorageFixture(browser, fixtureName);
-        return loginDialog.login(backEndMocks.getLoginDetails(fixtureName));
-    });
-
-    this.Given("I'm a CWL Instructor with students", {timeout: 20 * 1000}, function () {
-        var fixtureName = 'instructor/cwl_has_students_fixture';
-        backEndMocks.setStorageFixture(browser, fixtureName);
-        return loginDialog.skipLogin();
-    });
-
-    this.Given("I'm an Instructor with courses", {timeout: 20 * 1000}, function () {
-        var fixtureName = 'instructor/has_courses_fixture';
-        backEndMocks.setStorageFixture(browser, fixtureName);
-        return loginDialog.login(backEndMocks.getLoginDetails(fixtureName));
-    });
-
-    this.Given("I'm an Instructor with assignments", {timeout: 20 * 1000}, function () {
-        var fixtureName = 'instructor/has_assignments_fixture';
-        backEndMocks.setStorageFixture(browser, fixtureName);
-        return loginDialog.login(backEndMocks.getLoginDetails(fixtureName));
-    });
-};
-
-module.exports = setupInstructorStepDefinitionsWrapper;
