+++ acj/static/test/features/view_home.feature
@@ -2,7 +2,7 @@ Feature: View Home
   As user, I want to view courses on the home page
 
   Scenario: Loading home page as admin
-    Given I'm a System Administrator with courses
+    Given I'm a System Administrator
     And I'm on 'home' page
     Then I should see my courses with names:
       | name     | year | term    |
@@ -10,7 +10,7 @@ Feature: View Home
       | PHYS 101 | 2015 | Winter  |
 
   Scenario: Loading home page as instructor
-    Given I'm an Instructor with courses
+    Given I'm an Instructor
     And I'm on 'home' page
     Then I should see my courses with names:
       | name     | year | term    |
@@ -18,7 +18,7 @@ Feature: View Home
       | PHYS 101 | 2015 | Winter  |
 
   Scenario: Filtering home page courses as instructor
-    Given I'm an Instructor with courses
+    Given I'm an Instructor
     And I'm on 'home' page
     When I filter home page courses by 'CHEM'
     Then I should see '1' courses
@@ -27,7 +27,7 @@ Feature: View Home
       | CHEM 111 | 2015 | Winter  |
 
   Scenario: Loading home page as student
-    Given I'm a Student with courses
+    Given I'm a Student
     And I'm on 'home' page
     Then I should see my courses with names:
       | name     | year | term    |
