+++ acj/static/test/fixtures/instructor/default_fixture.js
@@ -4,9 +4,25 @@ var userFactory = new UserFactory();
 var SessionFactory  = require('../../factories/session_factory.js');
 var sessionFactory = new SessionFactory();
 
+var CourseFactory  = require('../../factories/course_factory.js');
+var courseFactory = new CourseFactory();
+
+var AssignmentFactory  = require('../../factories/assignment_factory.js');
+var assignmentFactory = new AssignmentFactory();
+
+var CriterionFactory  = require('../../factories/criterion_factory.js');
+var criterionFactory = new CriterionFactory();
+
 var storage = {
     session: {},
-    users: {}
+    users: {},
+    courses: {},
+    user_courses: {},
+    groups: [],
+    assignments: {},
+    course_assignments: {},
+    criteria: {},
+    user_search_results: {}
 }
 
 var admin = userFactory.generateUser("1abcABC123-abcABC123_Z", "System Administrator", {
@@ -28,10 +44,153 @@ var instructor = userFactory.generateUser("2abcABC123-abcABC123_Z", "Instructor"
 });
 storage.users[instructor.id] = instructor;
 
+var student1 = userFactory.generateUser("3abcABC123-abcABC123_Z", "Student", {
+    username: "student1",
+    displayname: "First Student",
+    firstname: "First",
+    lastname: "Student",
+    fullname: "First Student",
+    email: "first.student@exmple.com"
+});
+storage.users[student1.id] = student1;
 
+var student2 = userFactory.generateUser("4abcABC123-abcABC123_Z", "Student", {
+    username: "student2",
+    displayname: "Second Student",
+    firstname: "Second",
+    lastname: "Student",
+    fullname: "Second Student",
+    email: "second.student@exmple.com"
+});
+storage.users[student2.id] = student2;
+
+var course = courseFactory.generateCourse("1abcABC123-abcABC123_Z", {
+    name: "CHEM 111",
+    year: 2015,
+    term: "Winter",
+    description: "<p>CHEM 111 description<p>",
+});
+storage.courses[course.id] = course;
+
+var course2 = courseFactory.generateCourse("2abcABC123-abcABC123_Z", {
+    name: "PHYS 101",
+    year: 2015,
+    term: "Winter",
+    description: "<p>PHYS 101  description<p>",
+});
+storage.courses[course2.id] = course2;
+
+var group1 = "First Group";
+storage.groups.push(group1);
+var group2 = "Second Group";
+storage.groups.push(group2);
+var group3 = "Third Group";
+storage.groups.push(group3);
+
+
+var defaultCriterion = criterionFactory.getDefaultCriterion();
+storage.criteria[defaultCriterion.id] = defaultCriterion;
+
+var criterion2 = criterionFactory.generateCriterion("2abcABC123-abcABC123_Z", instructor.id, {
+    "name": "Which sounds better?",
+    "description": "<p>Choose the response that you think sounds more accurate of the two.</p>",
+    "default": true,
+});
+storage.criteria[criterion2.id] = criterion2;
+
+var criterion3 = criterionFactory.generateCriterion("3abcABC123-abcABC123_Z", instructor.id, {
+    "name": "Which looks better?",
+    "description": "<p>Choose the response that you think looks more accurate of the two.</p>",
+    "default": false,
+    "compared": true
+});
+storage.criteria[criterion3.id] = criterion3;
+
+
+// user_courses
+storage.user_courses[instructor.id] = [
+    { courseId: course.id, courseRole: "Instructor", groupName: null },
+    { courseId: course2.id, courseRole: "Instructor", groupName: null }
+];
+
+storage.user_courses[student1.id] = [
+    { courseId: course.id, courseRole: "Student", groupName: group1 }
+];
+
+storage.course_assignments[course.id] = [];
+
+var assignment_finished = assignmentFactory.generateAssignment("1abcABC123-abcABC123_Z", instructor, [defaultCriterion, criterion3], {
+    "name": "Assignment Finished",
+    "students_can_reply": true,
+    "available": true,
+    "compared": true,
+    "compare_period": false,
+    "after_comparing": true,
+    "answer_period": false,
+    "content": "<p>This assignment should already be completed</p>"
+});
+storage.assignments[assignment_finished.id] = assignment_finished;
+storage.course_assignments[course.id].push(assignment_finished.id);
+
+var assignment_being_compared = assignmentFactory.generateAssignment("2abcABC123-abcABC123_Z", instructor, [defaultCriterion, criterion3], {
+    "name": "Assignment Being Compared",
+    "students_can_reply": true,
+    "available": true,
+    "compared": true,
+    "compare_period": true,
+    "after_comparing": false,
+    "answer_period": false,
+    "content": "<p>This assignment should be compared right now</p>"
+});
+storage.assignments[assignment_being_compared.id] = assignment_being_compared;
+storage.course_assignments[course.id].push(assignment_being_compared.id);
+
+var assignment_being_answered = assignmentFactory.generateAssignment("3abcABC123-abcABC123_Z", instructor, [defaultCriterion, criterion3], {
+    "name": "Assignment Being Answered",
+    "students_can_reply": true,
+    "available": true,
+    "compared": false,
+    "compare_period": false,
+    "after_comparing": false,
+    "answer_period": true,
+    "content": "<p>This assignment should be answered right now</p>"
+});
+storage.assignments[assignment_being_answered.id] = assignment_being_answered;
+storage.course_assignments[course.id].push(assignment_being_answered.id);
+
+var assignment_upcoming = assignmentFactory.generateAssignment("4abcABC123-abcABC123_Z", instructor, [defaultCriterion, criterion3], {
+    "name": "Assignment Upcoming",
+    "students_can_reply": true,
+    "available": false,
+    "compared": false,
+    "compare_period": false,
+    "after_comparing": false,
+    "answer_period": false,
+    "content": "<p>This assignment should be coming in the future</p>"
+});
+storage.assignments[assignment_upcoming.id] = assignment_upcoming;
+storage.course_assignments[course.id].push(assignment_upcoming.id);
+
+// user_search_results
+storage.user_search_results.objects = [student2];
+storage.user_search_results.total = 1;
 
 storage.loginDetails = { id: instructor.id, username: instructor.username, password: "password" };
-storage.session = sessionFactory.generateSession(instructor.id, instructor.system_role, {});
-
+var session = sessionFactory.generateSession(instructor.id, instructor.system_role, {
+    "Course": {
+        "delete": {'1abcABC123-abcABC123_Z': false, '2abcABC123-abcABC123_Z': false},
+        "edit": {'1abcABC123-abcABC123_Z': true, '2abcABC123-abcABC123_Z': true},
+        "manage": {'1abcABC123-abcABC123_Z': false, '2abcABC123-abcABC123_Z': false},
+        "read": {'1abcABC123-abcABC123_Z': true, '2abcABC123-abcABC123_Z': true},
+    },
+    "Assignment": {
+        "create": {'1abcABC123-abcABC123_Z': true, '2abcABC123-abcABC123_Z': true, '3abcABC123-abcABC123_Z': true, '4abcABC123-abcABC123_Z': true},
+        "delete": {'1abcABC123-abcABC123_Z': true, '2abcABC123-abcABC123_Z': true, '3abcABC123-abcABC123_Z': true, '4abcABC123-abcABC123_Z': true},
+        "edit": {'1abcABC123-abcABC123_Z': true, '2abcABC123-abcABC123_Z': true, '3abcABC123-abcABC123_Z': true, '4abcABC123-abcABC123_Z': true},
+        "manage": {'1abcABC123-abcABC123_Z': true, '2abcABC123-abcABC123_Z': true, '3abcABC123-abcABC123_Z': true, '4abcABC123-abcABC123_Z': true},
+        "read": {'1abcABC123-abcABC123_Z': true, '2abcABC123-abcABC123_Z': true, '3abcABC123-abcABC123_Z': true, '4abcABC123-abcABC123_Z': true}
+    },
+});
+storage.session = session;
 
 module.exports = storage;
